Rearrange files to make compilation easier.
Pull code that requires template generation into files that
aren't compiled so often.
