+++ Cascade/WrapBoard.cc
@@ -1,0 +1,9 @@
+#include "WrapBoard.hh"
+#include "PrintStruct.hh"
+
+namespace cascade {
+
+DEFINEPRINT(MConfig)
+DEFINEPRINT(HSetup)
+
+}
