+++ Cascade/WrapFirmware.cc
@@ -1,0 +1,11 @@
+#include "WrapFirmware.hh"
+#include "PrintStruct.hh"
+
+namespace cascade {
+
+DEFINEPRINT(CFSimulation)
+DEFINEPRINT(CFStatistics)
+DEFINEPRINT(CFConfig)
+DEFINEPRINT(CFDiagnostics)
+
+}
