+++ Cascade/WrapFirmware.hh
@@ -102,4 +102,10 @@ BOOST_FUSION_ADAPT_STRUCT(
   (DWORD, dwHBDataLostCounter)
 )
 
+namespace cascade {
+  void Print(const CFConfig& conf);
+  void Print(const CFDiagnostics& conf);
+  void Print(const CFStatistics& conf);
+  void Print(const CFSimulation& conf);
+}
 #endif
